Moves reusable logic to separate util
