+++ site/_js/events.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,21 +21,7 @@ import './web-components/checkbox-group';
 import {RenderEventCard} from './render-event-card';
 import {loadMore} from './behaviors/load-more';
 import memoize from './utils/memoize';
-
-// Todo - abstract to separate file and refactor collection
-const startOfDay = new Date();
-startOfDay.setHours(0, 0, 0, 0);
-const isPastEvent = event => {
-  return new Date(event.date).getTime() < startOfDay.getTime();
-};
-//Todo - end
-
-// Todo - abstract to separate file and refactor collection
-const sortDesc = (a, b) =>
-  new Date(b.date).getTime() - new Date(a.date).getTime();
-const sortAsc = (a, b) =>
-  new Date(a.date).getTime() - new Date(b.date).getTime();
-//Todo - end
+import {isPastEvent, sortAsc, sortDesc} from './utils/events';
 
 const getEvents = async () => {
   const response = await fetch('/events.json');
@@ -44,17 +30,16 @@ const getEvents = async () => {
     throw new Error('Unable to fetch /events.json');
   }
 
-  const events = await response.json();
+  const all = await response.json();
 
   return {
-    upcomingEvents: events.filter(event => !isPastEvent(event)),
-    pastEvents: events.filter(event => isPastEvent(event)),
+    upcomingEvents: all.filter(event => !isPastEvent(event)),
+    pastEvents: all.filter(event => isPastEvent(event)),
   };
 };
 
 (async function () {
-  //todo:
-  //- ensure elements exist
+  //todo:ensure elements exist
   const getMemoizedEvents = memoize(getEvents);
 
   loadMore(
